+++ src/App.js
@@ -14,8 +14,8 @@ import NavBar from "./components/navbar";
 import Footer from "./components/footer";
 import Search from "./components/search.js";
 import DashboardGrid from "./components/dashboard/dashboard";
-import registration from "./components/registration";
-import Login from "./components/login";
+import registration from "./components/authentication/Registration";
+import Login from "./components/authentication/Login";
 import EditAccount from "./components/settings/editaccount";
 import CompSet from "./components/compSet";
 import AboutUs from "./components/aboutus";
