+++ src/components/authentication/Registration.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
-import { shouldUpdateLoggedInUser } from "../actions/index";
+import { shouldUpdateLoggedInUser } from "../../actions/index";
 
 import { makeStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
